refactor: divided responsabilities
