+++ src/features/register/components/DateOfBirth.tsx
@@ -1,4 +1,3 @@
-import { FormLabel } from '@/components/ui/form'
 import {
   Select,
   SelectContent,
@@ -7,43 +6,9 @@ import {
   SelectValue,
 } from '@/components/ui/select'
 import { useEffect, useState } from 'react'
-
-const Meses = [
-  { name: 'Janeiro', value: '01' },
-  { name: 'Fevereiro', value: '02' },
-  { name: 'Março', value: '03' },
-  { name: 'Abril', value: '04' },
-  { name: 'Maio', value: '05' },
-  { name: 'Junho', value: '06' },
-  { name: 'Julho', value: '07' },
-  { name: 'Agosto', value: '08' },
-  { name: 'Setembro', value: '09' },
-  { name: 'Outubro', value: '10' },
-  { name: 'Novembro', value: '11' },
-  { name: 'Dezembro', value: '12' },
-]
-
-function generateDayFromMonth() {
-  const days = []
-  // eslint-disable-next-line no-plusplus
-  for (let i = 1; i <= 31; i++) {
-    if (i < 10) {
-      days.push(`0${i}`)
-    } else days.push(i)
-  }
-  return days
-}
-
-function generateYearsUntillTodayFrom(oldestYear: number) {
-  const currentDate = new Date().getFullYear()
-  const maxOldDate = currentDate - oldestYear
-  const years = []
-  // eslint-disable-next-line no-plusplus
-  for (let i = currentDate; i > maxOldDate; i--) {
-    years.push(i)
-  }
-  return years
-}
+import { generateDaysOfAMonth } from '../helper/generateDaysOfAMonth'
+import { monthsArray } from '../helper/monthsArray'
+import { generateXYearsUntillToday } from '../helper/generateXYearsUntillToday'
 
 interface DateOfBirthProps {
   onSelectedDate: (date: string) => void
@@ -54,81 +19,67 @@ export function DateOfBirth({ onSelectedDate }: DateOfBirthProps) {
   const [monthForm, setMonthForm] = useState('')
   const [yearForm, setYearForm] = useState('')
 
-  // function handleSelectDate() {
-  //   onSelectedDate(`${monthForm}-${dayForm}-${yearForm}`)
-  // }
-
   useEffect(() => {
-    // handleSelectDate()
     if (dayForm !== '' && monthForm !== '' && yearForm !== '') {
       onSelectedDate(`${monthForm}-${dayForm}-${yearForm}`)
     }
-    // const handleSelectDate = () => {
-    //   if (dayForm !== '' && monthForm !== '' && yearForm !== '') {
-    //     onSelectedDate(`${monthForm}-${dayForm}-${yearForm}`)
-    //   }
-    // }
-    // handleSelectDate()
   }, [dayForm, monthForm, yearForm, onSelectedDate])
 
   return (
-    <>
-      <FormLabel>Data de nascimento*</FormLabel>
-      <div className='flex justify-between'>
-        <Select onValueChange={e => setDayForm(e)}>
-          <SelectTrigger className='w-16'>
-            <SelectValue placeholder='Dia' />
-          </SelectTrigger>
-          <SelectContent className='w-16'>
-            {generateDayFromMonth().map(day => {
-              return (
-                <SelectItem
-                  key={day}
-                  value={day.toString()}
-                >
-                  {day}
-                </SelectItem>
-              )
-            })}
-          </SelectContent>
-        </Select>
+    <div className='flex justify-between'>
+      <Select onValueChange={e => setDayForm(e)}>
+        <SelectTrigger className='w-16'>
+          <SelectValue placeholder='Dia' />
+        </SelectTrigger>
+        <SelectContent className='w-16'>
+          {generateDaysOfAMonth().map(day => {
+            return (
+              <SelectItem
+                key={day}
+                value={day.toString()}
+              >
+                {day}
+              </SelectItem>
+            )
+          })}
+        </SelectContent>
+      </Select>
 
-        <Select onValueChange={e => setMonthForm(e)}>
-          <SelectTrigger className='w-40'>
-            <SelectValue placeholder='Selecionar mes' />
-          </SelectTrigger>
-          <SelectContent>
-            {Meses.map(mes => {
-              return (
-                <SelectItem
-                  key={mes.value}
-                  value={mes.value.toString()}
-                >
-                  {mes.name}
-                </SelectItem>
-              )
-            })}
-          </SelectContent>
-        </Select>
+      <Select onValueChange={e => setMonthForm(e)}>
+        <SelectTrigger className='w-40'>
+          <SelectValue placeholder='Selecionar mes' />
+        </SelectTrigger>
+        <SelectContent>
+          {monthsArray.map(mes => {
+            return (
+              <SelectItem
+                key={mes.value}
+                value={mes.value.toString()}
+              >
+                {mes.name}
+              </SelectItem>
+            )
+          })}
+        </SelectContent>
+      </Select>
 
-        <Select onValueChange={e => setYearForm(e)}>
-          <SelectTrigger className='w-20'>
-            <SelectValue placeholder='Ano' />
-          </SelectTrigger>
-          <SelectContent>
-            {generateYearsUntillTodayFrom(100).map(ano => {
-              return (
-                <SelectItem
-                  key={ano}
-                  value={ano.toString()}
-                >
-                  {ano}
-                </SelectItem>
-              )
-            })}
-          </SelectContent>
-        </Select>
-      </div>
-    </>
+      <Select onValueChange={e => setYearForm(e)}>
+        <SelectTrigger className='w-20'>
+          <SelectValue placeholder='Ano' />
+        </SelectTrigger>
+        <SelectContent>
+          {generateXYearsUntillToday(100).map(ano => {
+            return (
+              <SelectItem
+                key={ano}
+                value={ano.toString()}
+              >
+                {ano}
+              </SelectItem>
+            )
+          })}
+        </SelectContent>
+      </Select>
+    </div>
   )
 }
