--- src/features/register/helper/checkTerm.ts
@@ -1,4 +1,0 @@
-export function checkTerm(username: string, wordArray: string[]): boolean {
-  const lowerCaseUser = username.toLowerCase()
-  return wordArray.some(word => word.toLowerCase().includes(lowerCaseUser))
-}
